Support for ANSI escape sequences to allow images
I nearly have the rough POC working.  I captured the text output of
https://wttr.in/:iterm2 since that URL currently serves a full HTML
page.  The translation to jQuery Terminal formatting string is working
and the `test` command shows that the base64 encoded png can be
rendered inline.

The `weather2` command gets the mocked iterm2 output and parses it into
the terminal format, but it doesn't display for some reason.  If I print
out the terminal formatted output, and check it in isolation it
appears to be a valid image.  Haven't figured out what the problem is
yet.